pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sga-readiness"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = []
authors = [
  { name = "Mark Thomas", email = "thomasmark@meta.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = ["pyyaml"]

[project.scripts]
sga-readiness = "sga_readiness.cli:cli"

[project.urls]
Documentation = "https://github.com/Mark Thomas/sga-readiness#readme"
Issues = "https://github.com/Mark Thomas/sga-readiness/issues"
Source = "https://github.com/Mark Thomas/sga-readiness"

[tool.hatch.version]
path = "src/sga_readiness/__about__.py"

[tool.hatch.envs.default]
python = "3.12"

[tool.hatch.envs.hatch-test]
default-args = ["tests"]

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.12"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/sga_readiness tests}"

[tool.coverage.run]
source_pkgs = ["sga_readiness", "tests"]
branch = true
parallel = true
omit = [
  "src/sga_readiness/__about__.py",
]

[tool.coverage.paths]
sga_readiness = ["src/sga_readiness", "*/sga-readiness/src/sga_readiness"]
tests = ["tests", "*/sga-readiness/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
